частичная функциональность интервалов

+++ app/src/main/java/diploma/work/magnum_opus/model/StudyMaterial.kt
@@ -10,7 +10,7 @@ package diploma.work.magnum_opus.model
  */
 data class StudyMaterial(
     val id: Long? = null,
-    val intervalsId: Int,
+    val intervalsId: Long,
     val title: String,
     val content: String,
     val createdAt: Long,
